Fix bug di comunicazione e pulizia codice e cartelle obsolete

+++ server/utils.c
@@ -21,7 +21,7 @@ bool check_draw(int match[3][3]) {
     for (int i = 0; i < 3; i++) {
         for (int j = 0; j < 3; j++) {
             if (match[i][j] == 0) {
-                return false; // Ci sono ancora celle vuote, quindi non è un pareggio
+                return false;
             }
         }
     }
@@ -31,7 +31,6 @@ bool check_draw(int match[3][3]) {
 void convert_json_to_buffer(char *json, generic_buffer *buffer){
     cJSON *json_obj = cJSON_Parse(json);
     if (json_obj == NULL) {
-        //printf("Errore JSON");
         return;
     }
     cJSON *sig_item = cJSON_GetObjectItem(json_obj, "sig");
@@ -141,7 +140,7 @@ cJSON* build_first_connection_message(int socket_id, match *match_list) {
 
     cJSON *content = cJSON_CreateObject();
     cJSON_AddNumberToObject(content, "socket_id", socket_id);
-    cJSON_AddItemToObject(content, "match_list", get_match_list(match_list));  // restituisce array o oggetto
+    cJSON_AddItemToObject(content, "match_list", get_match_list(match_list));
 
     cJSON_AddItemToObject(message, "content", content);
 
@@ -185,7 +184,12 @@ cJSON* build_message(int socket_fd, message_type_enum message_type, generic_buff
             cJSON_AddItemToObject(content, "draw_result", handle_draw(&buffer->handle_draw, mem.match_list));
         break;
         case SIG_DELETE_MATCH:
-            cJSON_AddItemToObject(content, "match_list", delete_match(buffer->delete_match.match_id, mem.match_list));
+            if(mem.match_list[buffer->delete_match.match_id].owner_id != socket_fd){
+                cJSON_AddStringToObject(content, "info", "not your match");
+            }
+            else{
+                cJSON_AddItemToObject(content, "match_list", delete_match(buffer->delete_match.match_id, mem.match_list));
+            }
         break;
     }
 
@@ -197,7 +201,6 @@ int send_message(int socket_fd, message_type_enum message_type, generic_buffer *
     cJSON *message = build_message(socket_fd, message_type, buffer);
     char *message_str = cJSON_PrintUnformatted(message);
     int sent = send(socket_fd, message_str, strlen(message_str), 0);
-    
     free(message_str);
     cJSON_Delete(message);
     return sent;
